Refactor code, added unit test for adding item
